agot-bg-website/requirements.txt
aioredis==1.3.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.1.2
Automat==0.8.0
beautifulsoup4==4.8.2
boto3==1.22.4
botocore==1.25.4
certifi==2019.6.16
cffi==1.13.2
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
defusedxml==0.6.0
Django==3.0.2
django-bootstrap4==1.1.1
django-debug-toolbar==2.2
django-environ==0.4.5
django-oauth-toolkit==1.3.2
django-prometheus==1.0.15
django-storages==1.12.3
djangorestframework==3.11.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jmespath==1.0.0
msgpack==0.6.2
nltk==3.4.5
oauthlib==3.0.2
prometheus-client==0.7.1
psycopg2-binary==2.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.2
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
rope==0.16.0
s3transfer==0.5.2
sentry-sdk==0.14.0
service-identity==18.1.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.5
sqlparse==0.3.0
Twisted==19.10.0
txaio==18.8.1
urllib3==1.25.4
whitenoise==5.0.1
zope.interface==4.7.1
